+++ src/domain/settings.rs
@@ -51,11 +51,11 @@ pub struct GameSaveData {
     pub player: PlayerSaveData,
     pub save_timestamp: f64,
     pub tick: u32,
-    pub seed: u64,
+    pub seed: u32,
 }
 
 impl GameSaveData {
-    pub fn new(player: PlayerSaveData, save_timestamp: f64, tick: u32, seed: u64) -> Self {
+    pub fn new(player: PlayerSaveData, save_timestamp: f64, tick: u32, seed: u32) -> Self {
         Self {
             player,
             save_timestamp,
